move visitor interfaces to its namespace
