+++ src/Visitor/Factory/DeserializationVisitorFactory.php
@@ -18,13 +18,9 @@ declare(strict_types=1);
  * limitations under the License.
  */
 
-namespace JMS\Serializer\VisitorFactory;
+namespace JMS\Serializer\Visitor\Factory;
 
-use JMS\Serializer\Accessor\AccessorStrategyInterface;
-use JMS\Serializer\DeserializationContext;
-use JMS\Serializer\DeserializationVisitorInterface;
-use JMS\Serializer\GraphNavigatorInterface;
-use JMS\Serializer\SerializationContext;
+use JMS\Serializer\Visitor\DeserializationVisitorInterface;
 
 /**
  *
