+++ src/Visitor/Factory/JsonDeserializationVisitorFactory.php
@@ -18,13 +18,10 @@ declare(strict_types=1);
  * limitations under the License.
  */
 
-namespace JMS\Serializer\VisitorFactory;
+namespace JMS\Serializer\Visitor\Factory;
 
-use JMS\Serializer\Accessor\AccessorStrategyInterface;
-use JMS\Serializer\DeserializationContext;
-use JMS\Serializer\DeserializationVisitorInterface;
-use JMS\Serializer\GraphNavigatorInterface;
 use JMS\Serializer\JsonDeserializationVisitor;
+use JMS\Serializer\Visitor\DeserializationVisitorInterface;
 
 /**
  *
