+++ src/Visitor/Factory/XmlSerializationVisitorFactory.php
@@ -18,13 +18,9 @@ declare(strict_types=1);
  * limitations under the License.
  */
 
-namespace JMS\Serializer\VisitorFactory;
+namespace JMS\Serializer\Visitor\Factory;
 
-use JMS\Serializer\Accessor\AccessorStrategyInterface;
-use JMS\Serializer\GraphNavigatorInterface;
-use JMS\Serializer\JsonSerializationVisitor;
-use JMS\Serializer\SerializationContext;
-use JMS\Serializer\SerializationVisitorInterface;
+use JMS\Serializer\Visitor\SerializationVisitorInterface;
 use JMS\Serializer\XmlSerializationVisitor;
 
 /**
@@ -50,26 +46,26 @@ final class XmlSerializationVisitorFactory implements SerializationVisitorFactor
         );
     }
 
-    public function setDefaultRootName(string $name, ?string $namespace = null):self
+    public function setDefaultRootName(string $name, ?string $namespace = null): self
     {
         $this->defaultRootName = $name;
         $this->defaultRootNamespace = $namespace;
         return $this;
     }
 
-    public function setDefaultVersion(string $version):self
+    public function setDefaultVersion(string $version): self
     {
         $this->defaultVersion = $version;
         return $this;
     }
 
-    public function setDefaultEncoding(string $encoding):self
+    public function setDefaultEncoding(string $encoding): self
     {
         $this->defaultEncoding = $encoding;
         return $this;
     }
 
-    public function setFormatOutput(bool $formatOutput):self
+    public function setFormatOutput(bool $formatOutput): self
     {
         $this->formatOutput = (boolean)$formatOutput;
         return $this;
