+++ src/VisitorInterface.php
@@ -29,8 +29,7 @@ use JMS\Serializer\Metadata\PropertyMetadata;
  * This contains the minimal set of values that must be supported for any
  * output format.
  *
- * @internal Use SerializationVisitorInterface and DeserializationVisitorInterface instead
- *
+ * @internal
  * @author Johannes M. Schmitt <schmittjoh@gmail.com>
  * @author Asmir Mustafic <goetas@gmail.com>
  */
