+++ tests/Fixtures/InitializedObjectConstructor.php
@@ -22,7 +22,7 @@ namespace JMS\Serializer\Tests\Fixtures;
 
 use JMS\Serializer\Construction\ObjectConstructorInterface;
 use JMS\Serializer\DeserializationContext;
-use JMS\Serializer\DeserializationVisitorInterface;
+use JMS\Serializer\Visitor\DeserializationVisitorInterface;
 use JMS\Serializer\Metadata\ClassMetadata;
 
 /**
