+++ tests/Serializer/GraphNavigatorTest.php
@@ -25,7 +25,7 @@ use JMS\Serializer\Accessor\DefaultAccessorStrategy;
 use JMS\Serializer\Construction\UnserializeObjectConstructor;
 use JMS\Serializer\DeserializationContext;
 use JMS\Serializer\GraphNavigator\DeserializationGraphNavigator;
-use JMS\Serializer\DeserializationVisitorInterface;
+use JMS\Serializer\Visitor\DeserializationVisitorInterface;
 use JMS\Serializer\EventDispatcher\EventDispatcher;
 use JMS\Serializer\Exclusion\ExclusionStrategyInterface;
 use JMS\Serializer\GraphNavigatorInterface;
@@ -35,7 +35,8 @@ use JMS\Serializer\Metadata\Driver\AnnotationDriver;
 use JMS\Serializer\Naming\IdenticalPropertyNamingStrategy;
 use JMS\Serializer\SerializationContext;
 use JMS\Serializer\GraphNavigator\SerializationGraphNavigator;
-use JMS\Serializer\SerializationVisitorInterface;
+use JMS\Serializer\Visitor\SeerializationVisitorInterface;
+use JMS\Serializer\Visitor\SerializationVisitorInterface;
 use Metadata\MetadataFactory;
 
 class GraphNavigatorTest extends \PHPUnit\Framework\TestCase
